Move ui events into respective packages
