+++ src/main/java/seedu/restaurant/commons/events/ui/ingredient/DisplayIngredientListRequestEvent.java
@@ -1,4 +1,4 @@
-package seedu.restaurant.commons.events.ui;
+package seedu.restaurant.commons.events.ui.ingredient;
 
 import seedu.restaurant.commons.events.BaseEvent;
