+++ src/main/java/seedu/restaurant/logic/commands/ingredient/EditIngredientByNameCommand.java
@@ -5,7 +5,7 @@ import static seedu.restaurant.model.Model.PREDICATE_SHOW_ALL_INGREDIENTS;
 
 import seedu.restaurant.commons.core.EventsCenter;
 import seedu.restaurant.commons.core.Messages;
-import seedu.restaurant.commons.events.ui.DisplayIngredientListRequestEvent;
+import seedu.restaurant.commons.events.ui.ingredient.DisplayIngredientListRequestEvent;
 import seedu.restaurant.logic.CommandHistory;
 import seedu.restaurant.logic.commands.CommandResult;
 import seedu.restaurant.logic.commands.exceptions.CommandException;
