+++ src/main/java/seedu/restaurant/logic/commands/ingredient/ListIngredientsCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 import static seedu.restaurant.model.Model.PREDICATE_SHOW_ALL_INGREDIENTS;
 
 import seedu.restaurant.commons.core.EventsCenter;
-import seedu.restaurant.commons.events.ui.DisplayIngredientListRequestEvent;
+import seedu.restaurant.commons.events.ui.ingredient.DisplayIngredientListRequestEvent;
 import seedu.restaurant.logic.CommandHistory;
 import seedu.restaurant.logic.commands.Command;
 import seedu.restaurant.logic.commands.CommandResult;
