+++ src/main/java/seedu/restaurant/ui/MainWindow.java
@@ -19,19 +19,19 @@ import javafx.stage.Stage;
 import seedu.restaurant.commons.core.Config;
 import seedu.restaurant.commons.core.GuiSettings;
 import seedu.restaurant.commons.core.LogsCenter;
-import seedu.restaurant.commons.events.ui.DisplayIngredientListRequestEvent;
-import seedu.restaurant.commons.events.ui.DisplayRecordListRequestEvent;
-import seedu.restaurant.commons.events.ui.DisplaySalesReportEvent;
 import seedu.restaurant.commons.events.ui.ExitAppRequestEvent;
-import seedu.restaurant.commons.events.ui.LoginEvent;
-import seedu.restaurant.commons.events.ui.LogoutEvent;
-import seedu.restaurant.commons.events.ui.RecordPanelSelectionChangedEvent;
 import seedu.restaurant.commons.events.ui.ShowHelpRequestEvent;
 import seedu.restaurant.commons.events.ui.accounts.DisplayAccountListRequestEvent;
+import seedu.restaurant.commons.events.ui.accounts.LoginEvent;
+import seedu.restaurant.commons.events.ui.accounts.LogoutEvent;
+import seedu.restaurant.commons.events.ui.ingredient.DisplayIngredientListRequestEvent;
 import seedu.restaurant.commons.events.ui.menu.DisplayItemListRequestEvent;
 import seedu.restaurant.commons.events.ui.menu.ItemPanelSelectionChangedEvent;
 import seedu.restaurant.commons.events.ui.reservation.DisplayReservationListRequestEvent;
 import seedu.restaurant.commons.events.ui.reservation.ReservationPanelSelectionChangedEvent;
+import seedu.restaurant.commons.events.ui.sales.DisplayRecordListRequestEvent;
+import seedu.restaurant.commons.events.ui.sales.DisplaySalesReportEvent;
+import seedu.restaurant.commons.events.ui.sales.RecordPanelSelectionChangedEvent;
 import seedu.restaurant.logic.Logic;
 import seedu.restaurant.model.UserPrefs;
 import seedu.restaurant.ui.account.AccountListPanel;
