+++ src/main/java/seedu/restaurant/ui/sales/RecordListPanel.java
@@ -9,7 +9,7 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.restaurant.commons.core.LogsCenter;
-import seedu.restaurant.commons.events.ui.RecordPanelSelectionChangedEvent;
+import seedu.restaurant.commons.events.ui.sales.RecordPanelSelectionChangedEvent;
 import seedu.restaurant.model.salesrecord.SalesRecord;
 import seedu.restaurant.ui.UiPart;
