+++ src/test/java/seedu/restaurant/commons/events/ui/DisplayRecordListRequestEventTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 
 import seedu.restaurant.commons.events.BaseEvent;
+import seedu.restaurant.commons.events.ui.sales.DisplayRecordListRequestEvent;
 
 public class DisplayRecordListRequestEventTest {
     @Test
