+++ src/test/java/seedu/restaurant/ui/account/UsernameDisplayTest.java
@@ -7,8 +7,8 @@ import org.junit.Before;
 import org.junit.Test;
 
 import guitests.guihandles.accounts.UsernameDisplayHandle;
-import seedu.restaurant.commons.events.ui.LoginEvent;
-import seedu.restaurant.commons.events.ui.LogoutEvent;
+import seedu.restaurant.commons.events.ui.accounts.LoginEvent;
+import seedu.restaurant.commons.events.ui.accounts.LogoutEvent;
 import seedu.restaurant.testutil.account.AccountBuilder;
 import seedu.restaurant.ui.GuiUnitTest;
