+++ src/test/java/systemtests/ClearCommandSystemTest.java
@@ -8,7 +8,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import seedu.restaurant.commons.core.index.Index;
-import seedu.restaurant.commons.events.ui.LoginEvent;
+import seedu.restaurant.commons.events.ui.accounts.LoginEvent;
 import seedu.restaurant.logic.commands.ClearCommand;
 import seedu.restaurant.logic.commands.RedoCommand;
 import seedu.restaurant.logic.commands.UndoCommand;
