+++ src/test/java/systemtests/FindCommandSystemTest.java
@@ -16,7 +16,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import seedu.restaurant.commons.core.index.Index;
-import seedu.restaurant.commons.events.ui.LoginEvent;
+import seedu.restaurant.commons.events.ui.accounts.LoginEvent;
 import seedu.restaurant.logic.commands.DeleteCommand;
 import seedu.restaurant.logic.commands.FindCommand;
 import seedu.restaurant.logic.commands.RedoCommand;
