+++ src/test/java/systemtests/SelectCommandSystemTest.java
@@ -15,7 +15,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import seedu.restaurant.commons.core.index.Index;
-import seedu.restaurant.commons.events.ui.LoginEvent;
+import seedu.restaurant.commons.events.ui.accounts.LoginEvent;
 import seedu.restaurant.logic.commands.RedoCommand;
 import seedu.restaurant.logic.commands.SelectCommand;
 import seedu.restaurant.logic.commands.UndoCommand;
